backend/requirements.txt
pydantic==1.10.7
fastapi==0.95.0
uvicorn[standard]==0.21.1
websockets==11.0.3
python-dotenv==1.0.0
supabase>=2.3.0
google-generativeai==0.3.0
slowapi==0.1.8 
python-multipart==0.0.6